Many changes and fixes
Added the shebang and encoding to the top of the file.

Moved the copyright up in the file and commented it differently.

Removed many redundant comments throughout the file.

Added missing docstrings.

Changed how the font is set, for cross platform compatibility.

Separated some methods into two parts; an 'event' method and an 'action'
method.

Gracefully handles an operator trying to do things like save or close
when they do not have any file open or have the notebook hidden.

Resize layout properly when showing or hiding things like the notebook
or status bar.

Rewritten logic in some places to reduce code.

Now able to open multiple files at once (some remaining issues to be
resolved here).

Added comments to explain what happens when a notebook page is deleted.

Removed printing method because it was not cross platform compatible.

Created a method for notification dialogs.